shork bot/requirements.txt
discord.py
dnspython
motor~=2.4.0
aiohttp~=3.7.4.post0
EZPaginator~=1.3.4
mathterpreter~=1.1.2

pymongo~=3.11.4
dataclasses~=0.6